requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
ollama>=0.1.4
faiss-cpu==1.7.4
numpy>=1.24.0
python-magic>=0.4.27
tqdm>=4.65.0
aiofiles>=23.2.1
